+++ modules/utils/getByteLength.js
@@ -1,8 +1,0 @@
-function getByteLength(chunk) {
-  if (typeof chunk.byteLength !== 'undefined')
-    return chunk.byteLength;
-
-  return chunk.length;
-}
-
-module.exports = getByteLength;